Move TagService to TimeScout.Application
